+++ cspell.json
@@ -20,29 +20,50 @@
     ],
     "language": "en-US",
     "words": [
+        "blatt",
+        "breit",
+        "clebsch",
         "conda",
         "docstrings",
         "expertsystem",
+        "gordan",
         "graphviz",
         "helicity",
+        "ipynb",
+        "ipython",
         "isospin",
         "jupyter",
+        "mypy",
         "numpy",
         "pbar",
-        "mypy",
+        "pycompwa",
         "pydocstyle",
         "pylint",
         "pytest",
         "redeboer",
         "spflueger",
         "tensorwaves",
-        "venv"
+        "venv",
+        "weisskopf",
+        "xmltodict"
     ],
     "ignoreWords": [
+        "celltoolbar",
         "codacy",
+        "codemirror",
+        "cparity",
+        "kernelspec",
+        "mimetype",
+        "nbconvert",
+        "nbformat",
         "nbsphinx",
+        "phsp",
+        "pygments",
         "pyproject",
+        "rightarrow",
         "rtfd",
+        "stm's",
+        "unparse",
         "untracked"
     ]
 }
