+++ tests/amplitude/test_yaml_helicity.py
@@ -6,7 +6,7 @@ import pytest
 
 import yaml
 
-from expertsystem.amplitude.helicitydecay import HelicityAmplitudeGenerator
+from expertsystem.amplitude.helicity_decay import HelicityAmplitudeGenerator
 from expertsystem.ui.system_control import (
     InteractionTypes,
     StateTransitionManager,
