+++ tests/channels/test_y_to_d0d0barpi0pi0.py
@@ -7,7 +7,7 @@ import logging
 import pytest
 
 from expertsystem.amplitude.canonicaldecay import CanonicalAmplitudeGenerator
-from expertsystem.amplitude.helicitydecay import HelicityAmplitudeGenerator
+from expertsystem.amplitude.helicity_decay import HelicityAmplitudeGenerator
 from expertsystem.state.particle import (
     InteractionQuantumNumberNames,
     create_spin_domain,
